+++ persons.json
@@ -1,0 +1,1 @@
+[{"json_class":"Student","id":546,"name":"amalia","age":23,"parent_permission":true},{"json_class":"Teacher","id":786,"name":"egdafgb","age":23,"specialization":"asfsds"}]
