+++ rentals.json
@@ -1,0 +1,1 @@
+[{"json_class":"Rental","date":"04/05/2023","book":"huckleberry finn adventures","person":603}]
